Add files via upload
An interactive GUI program that tells us whether a star will turn out to be a black hole or a dwarf star at the end of its life cycle. This is based on a scientific limit known as the Chandrashekar limit. If our input in terms of solar masses is lesser than 1.4, then a star would likely end up as a dwarf star. Also, its radius and potential energy are displayed as a result. If not, then the star would end up as a black hole. The type of black hole is displayed based on the input of mass, followed by information on whether or not we could escape a black hole based on input parameters such as escape velocity and the angle.